refactor: make toplevel makefile work
